Translate client block to starknet_api block (#201)
* Translate client block to starknet_api block

* Fix CR

+++ crates/starknet_api/src/transaction.rs
@@ -16,7 +16,7 @@ pub struct TransactionHash(pub StarkHash);
 #[derive(
     Debug, Default, Copy, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord,
 )]
-pub struct TransactionOffsetInBlock(pub u64);
+pub struct TransactionOffsetInBlock(pub usize);
 
 /// A fee in StarkNet.
 #[derive(
